+++ client/src/Color.cpp
@@ -1,6 +1,6 @@
 #include "Color.hh"
 
-Color::Color(uint32_t red, uint32_t green, uint32_t blue, uint32_t alpha, bool used) :
+Color::Color(uint8_t red, uint8_t green, uint8_t blue, uint8_t alpha, bool used) :
 	_r(red), _g(green), _b(blue), _a(alpha), _used(used)
 {
 }
